+++ FS22_BeesRevamp/src/brutils.lua
@@ -8,7 +8,7 @@
 --
 BrUtils = {
     DEBUG_MODE = false,
-    MOD_NAME = g_currentModName or "unknown",
+    MOD_NAME = g_currentModName or 'unknown',
     SEVERITY = {
         INFO = 1,
         ERROR = 2,
@@ -41,7 +41,7 @@ function BrUtils:getSeverityString(serverity)
         end
     end
 
-    return "{Severity not found}"
+    return '{Severity not found}'
 end
 
 ---Log a debug message
@@ -72,17 +72,30 @@ function BrUtils:logWarning(messageFormat, ...)
     self:log(BrUtils.SEVERITY.WARNING, messageFormat, ...)
 end
 
----TODO
+---BrUtils:getCurrentDayYearString
 ---@return string
 function BrUtils:getCurrentDayYearString()
-    return 'Y' .. g_currentMission.environment.currentYear .. 'M' .. g_currentMission.environment.currentPeriod .. 'D0'
+    return 'Y' .. g_currentMission.environment.currentYear .. 'M' .. self:getStockPeriod() .. 'D0'
 end
 
----TODO
+---BrUtils:getModText
 ---@param text string
 ---@return string
 function BrUtils:getModText(text)
     return g_i18n:getText(text, self.MOD_NAME)
+end
+
+---Get the current period. If TerraLife is enabled, translate the
+---weeks into the period.
+---@return number
+function BrUtils:getStockPeriod()
+    local period = g_currentMission.environment.currentPeriod
+
+    if g_modIsLoaded['FS22_TerraLifePlus'] then
+        period = math.floor(period * 0.25) + 1
+    end
+
+    return period
 end
 
 g_brUtils = BrUtils;