+++ CHANGELOG.md
@@ -4,3 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
+- Reduced load while playing with precision farming and calculating the field info
+
+## [1.0.0.0] - 2024-10-01
+- Initial Release
+
+## [0.0.3.0] - 2024-08-24
+- english translation of moddesc and helpline - for [#2](https://github.com/Peppie84/FS22_BeesRevamp/issues/2)
+- added overpopulation info to the field info - for [#1](https://github.com/Peppie84/FS22_BeesRevamp/issues/1)
+- fixed a bug with precision farming mod and showing not correct field infos
+
+## [0.0.2.1] - 2024-08-21
+- fix a bug for saving the `monthlyPressureCheck`
+
+## [0.0.2.0] - 2024-08-02
+- adjust the sell price of honey from 10.20 to 5.42
+- added the ability to `decideToSwarm` when played with multiple days per period. One decision per month
+- no nectar to honey production within the winter period if still nectar is in the hive
+- increased the bee nectar consumption from 0.000433 to 0.000453 per month
+- fixed the nectar/honey production and consumption when played with multiple days per period.
+- the nectar value on the info hud is now displayed in float,2